// SocialAnalizeF/SocialAnalize/src/app/models/chart/chart.model.ts
export class ChartModel {
    showXAxis: boolean;
    showYAxis: boolean;
    gradient: boolean;
    showLegend: boolean;
    showXAxisLabel: boolean;
    xAxisLabel: string;
    showYAxisLabel: boolean;
    yAxisLabel: string;
    autoScale: boolean;
    view: any[];
    colorScheme: any;
    data: any[];
}